% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMovies.R
\name{generateMoviesDataSetFromSourceData}
\alias{generateMoviesDataSetFromSourceData}
\title{Generation of the movie data set}
\usage{
generateMoviesDataSetFromSourceData()
}
\description{
Generation of the movie data set
}
