% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomMovies.R
\name{getRandomMovies}
\alias{getRandomMovies}
\title{Get a random subset of the movie data set with optional seed.}
\usage{
getRandomMovies(seed)
}
\arguments{
\item{seed, }{e.g "b1234567" or 18}
}
\value{
A data frame with a random numer of rows
}
\description{
This function creates a subset (of random rows) of the IMDB movie
data frame. Given a student id, this function will always return the
same subset.
}
\examples{
##Given the same student ID
##the same data set is returned
d1 = getRandomMovies("b1234567")
d2 = getRandomMovies("b1234567")
identical(d1, d2)
##Different IDs get different data
d3 = getRandomMovies("b1234568")
identical(d1, d3)
}
