% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music.R
\docType{data}
\name{music}
\alias{music}
\title{Music information for a selection of popular music from 1950 to 2009}
\format{An object of class \code{data.frame} with 424 rows and 10 columns.}
\usage{
data(music)
}
\value{
A data frame with 424 rows and 10 variables.
}
\description{
Music information for a selection of popular music from 1950 to 2009
}
\note{
The data set contains the following fields:
\describe{
\item{artist}{Artist name.}
\item{track}{Track name.}
\item{year}{Year of release.}
\item{decade}{Decade of release.}
\item{duration}{Track length in seconds.}
\item{tempo}{Tempo of the song in BPM (beats per minute)}
\item{energy}{A number that ranges from 0 to 1, representing how energetic the track is.}
\item{danceability}{A number that ranges from 0 to 1, representing how suitable a track is for dancing, using a number of musical elements. The combination of musical elements that best characterize danceability include tempo, rhythm stability, beat strength, and overall regularity.}
\item{key}{The musical key that the track is in.}
\item{genre}{A list of comma separated styles or categories taht the track is in.}
}
}
\examples{
data(music)
}
\references{
https://tsort.info/music, https://github.com/nvempala/Music-Decades-Top-100, https://www.discogs.com/developers/
}
\keyword{datasets}
