% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{The Complete Pokemon Dataset - more than 800 Pokemon}
\format{An object of class \code{data.frame} with 801 rows and 40 columns.}
\usage{
data(pokemon)
}
\value{
A data frame with 801 rows and 40 variables.
}
\description{
This dataset contains information on 801 Pokemon from the first Seven
Generations of Pokemon. The information contained in this dataset include
Base Stats, Performance against Other Types, Height, Weight, Classification,
Egg Steps, Experience Points, Abilities, etc. The information was scraped
from http://serebii.net/
}
\note{
The data set contains the following fields:
\describe{
\item{name}{The English name of the Pokemon}
\item{pokedex_number}{The entry number of the Pokemon in the National Pokedex}
\item{generation}{The numbered generation which the Pokemon was first introduced}
\item{type1}{The Primary Type of the Pokemon}
\item{type2}{The Secondary Type of the Pokemon}
\item{classification}{The Classification of the Pokemon as described by the Sun and Moon Pokedex}
\item{hp}{The Base HP of the Pokemon}
\item{attack}{The Base Attack of the Pokemon}
\item{defense}{The Base Defense of the Pokemon}
\item{sp_attack}{The Base Special Attack of the Pokemon}
\item{sp_defense}{The Base Special Defense of the Pokemon}
\item{speed}{The Base Speed of the Pokemon}
\item{is_legendary}{Denotes if the Pokemon is legendary}
\item{japanese_name}{The Original Japanese name of the Pokemon}
\item{percentage_male}{The percentage of the species that are male}
\item{height_m}{Height of the Pokemon in metres}
\item{weight_kg}{The Weight of the Pokemon in kilograms}
\item{capture_rate}{Capture Rate of the Pokemon}
\item{base_egg_steps}{The number of steps required to hatch an egg of the Pokemon}
\item{abilities}{A list of abilities that the Pokemon is capable of having}
\item{experience_growth}{The Experience Growth of the Pokemon}
\item{base_happiness}{Base Happiness of the Pokemon}
\item{against_N}{Eighteen features that denote the amount of damage taken against an attack of a particular type N}}
}
\examples{
data(pokemon)
}
\references{
The information was scraped from http://serebii.net/ and licensed CC0: Public Domain
}
\keyword{datasets}
