% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\docType{data}
\name{tv}
\alias{tv}
\title{TV show information and user ratings from imdb.com.}
\format{An object of class \code{data.frame} with 1562 rows and 22 columns.}
\usage{
data(tv)
}
\value{
A data frame with 1562 rows and 22 variables.
}
\description{
The Internet Movie Database, \url{http://imdb.com/}, is a
website devoted to collecting movie data supplied by studios and fans.
It claims to be the biggest movie database on the web and is run by
Amazon.
}
\details{
TV shows were selected for inclusion if they were marked as "not adult",
and had at least 5000 rating votes.
}
\note{
The data set contains the following fields:
\describe{
\item{Title}{Title of the tv show}
\item{StartYear}{Year of release.}
\item{EndYear}{Year the show ended (if known).}
\item{Type}{Type of show (tvMiniSeries,tvSeries,tvShort,tvSpecial).}
\item{Rating}{Average IMDb user rating.}
\item{Votes}{Number of IMDb users who rated this movie.}
\item{Length}{Length in minutes.}
\item{Action, Animation, Comedy, Drama, Documentary, Fantasy, Horror, Romance, SciFi, Short}{Binary variables representing if movie was classified as belonging to that genre.}}
}
\examples{
data(tv)
}
\references{
Information courtesy of IMDb (http://www.imdb.com). IMDb data files are publically available at https://datasets.imdbws.com/
}
\keyword{datasets}
